add geoexplorer to vis
