+++ package.json
@@ -15,8 +15,8 @@
         "@mui/icons-material": "^6.0.2",
         "@mui/joy": "^5.0.0-beta.49",
         "@mui/material": "^6.0.2",
-        "@ncsa/geo-explorer": "^0.1.0",
         "@mui/x-data-grid": "^7.28.3",
+        "@ncsa/geo-explorer": "^0.1.0",
         "@reduxjs/toolkit": "^2.2.7",
         "@rjsf/mui": "^5.24.3",
         "@rjsf/utils": "^5.24.3",
@@ -25,6 +25,8 @@
         "@types/react-gravatar": "^2.6.14",
         "@xyflow/react": "^12.2.1",
         "axios": "~1.7",
+        "geostyler-sld-parser": "^7.0.0",
+        "geostyler-style": "^7.0.0",
         "i": "^0.3.7",
         "maplibre-gl": "^5.2.0",
         "npm": "^11.2.0",
