+++ src/components/Project/LegacyProjectPage.tsx
@@ -36,7 +36,7 @@ import DFR3Icon from "@mui/icons-material/ShowChart";
 import HazardIcon from "@mui/icons-material/Storm";
 import VisualizationIcon from "@mui/icons-material/Map";
 import { CreateVisualizationDialog } from "@app/components/Project/Resource/VisualizationDialog";
-import { VisualizationView } from "@app/components/Project/Resource/VisaualizationView";
+import { VisualizationView } from "@app/components/Project/Resource/VisualizationView";
 import Snackbar from "@mui/joy/Snackbar";
 import { AddFromServiceDialog } from "@app/components/Project/Resource/AddFromServiceDialog";
 import { CreateWorkflowDialog } from "./CreateWorkflow";
